Update Wayland packages (#3111)
