+++ packages/xorg_server.rb
@@ -28,7 +28,7 @@ class Xorg_server < Package
   depends_on 'libtirpc'
   depends_on 'font_util'
   depends_on 'libunwind'
-  depends_on 'wayland'
+  depends_on 'wayland_protocols'
   depends_on 'nettle'
   depends_on 'libbsd'
   depends_on 'dbus'
